config/requirements-minimal.txt
# Trading Bot v10 Enterprise - Dependencias Mínimas
# =================================================
# Solo las dependencias esenciales para funcionamiento básico

# ==================== CORE ESSENTIAL ====================
# Core dependencies mínimas
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# ==================== TRADING ESSENTIAL ====================
# Exchange API esencial
ccxt>=4.0.0

# WebSocket esencial
websockets>=11.0.0

# ==================== ASYNC ESSENTIAL ====================
# Async support esencial
aiofiles>=23.0.0
tenacity>=8.2.0

# ==================== CONFIGURATION ESSENTIAL ====================
# Configuración esencial
PyYAML>=6.0
python-dotenv>=1.0.0

# ==================== LOGGING ESSENTIAL ====================
# Logging esencial
colorlog>=6.7.0

# ==================== UTILITIES ESSENTIAL ====================
# Utilidades esenciales
tqdm>=4.65.0
click>=8.1.0

# ==================== INSTALLATION ====================
# Para instalación mínima:
# pip install -r requirements-minimal.txt
#
# Esto instalará solo las dependencias esenciales
# para funcionamiento básico del sistema
